spent 30mins tweaking behaviour for MoveStrategy + in range of poi's belifs to find the right balance to get them to work.
added Complete bool to IEntityInteractStrategies interface to notify when to InteractStrategy when an interaction is complete
(ended up using timer.IsFinishes bool on timers to achieve the same thing)

updated sensors to not return a capturable poi if something it already capturing it.

spent roughly an hr setting up beliefs, actions and goals for getting worker brain to pick up resources from one capture point then drop them off at a home base. works incredibbly well.

TODO:
filter out home bases from being a friendly poi in EntitySensors. (worker gets stuck picking/dropping off res at home base once there)
see where i can simplify/refine the beliefs, actions and goals.
continue testing how well it handles multiple pois that are neutral, friend enemy all in its range.

+++ Assets/Scripts/Entity Scripts/EntityStats.cs
@@ -10,7 +10,7 @@ public class EntityStats : MonoBehaviour
 	public EntityData _Data;
 	public float currentHealth;
 
-	public float carriedResources;
+	public int carriedResources;
 
 	CountdownTimer statsTimer;
